requirements-vercel.txt
# Core FastAPI dependencies
fastapi==0.114.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9

# HTTP client
requests==2.32.3

# Data processing
pandas==2.2.2
numpy==2.0.2

# PDF processing
pypdf==4.3.1
PyPDF2==3.0.1

# Environment management
python-dotenv==1.0.0

# Database support
sqlalchemy==2.0.25
alembic==1.13.1

# Enhanced file processing
python-magic==0.4.27
chardet==5.2.0

# Production dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Vercel specific
vercel==0.5.0
